backend/requirements.txt
fastapi
uvicorn
groq
requests
beautifulsoup4
python-dotenv
google-search-results
pydantic
email-validator
